+++ tests/Extractor/fixtures/formatter.php
@@ -7,11 +7,14 @@
 declare(strict_types=1);
 
 use FormatPHP\DescriptorCollection;
+use FormatPHP\Extractor\MessageExtractorOptions;
 
 /**
  * @return array<string, array{translation: string}>
+ *
+ * @psalm-suppress UnusedClosureParam
  */
-return function (DescriptorCollection $collection): array {
+return function (DescriptorCollection $collection, MessageExtractorOptions $options): array {
     $format = [];
     foreach ($collection as $item) {
         $format[(string) $item->getId()] = [